requirements.txt
flask==1.0.2
tweepy==3.6.0
python-decouple==3.1
pytest=3.7.4